Updated controller to print out Gyro x and y
and i had to change GetPadded to PrintPadded

For some reson that i dont YET understand when i tried to print two values it would print some random memory
What was wered was that i only changed it in 'I' and it happened to both 'i' (RAW) (unedited) and 'I' (Normalized)(edited). IDK what happened.
But its fine if i printed it directly in GetPaddedValue, so its now PrintPaddedValue :)...
Strange